Fix roundToInt signature

+++ Source/WebCore/platform_scion/graphics/LayoutPoint.swift
@@ -177,7 +177,7 @@ func ceiledLayoutPoint(p: FloatPoint) -> LayoutPointWrapper {
 func snappedIntSize(_ size: LayoutSizeWrapper, _ location: LayoutPointWrapper) -> IntSize {
   let snap = { (_ a: LayoutUnit, _ b: LayoutUnit) -> Int32 in
     let fraction = b.fraction()
-    return Int32(roundToInt(value: fraction + a)) - Int32(roundToInt(value: fraction))
+    return roundToInt(value: fraction + a) - roundToInt(value: fraction)
   }
   return IntSize(width: snap(size.width(), location.x), height: snap(size.height(), location.y))
 }
